Fixed a bug in reading mtx-files in float-format and added a mode for running with maximum posterior gene-variance

+++ src/ReadInputFiles.cpp
@@ -55,7 +55,7 @@ void Get_G_C_UMIcountMatrix(string in_file, int &N_rows, int &G, int &C, int N_c
             /****go to the next in the list of words****/
             token = strtok(NULL," \t,");
             if(token != NULL){
-                n += atoi(token);/**total count this gene**/
+                n += stod(token);/**total count this gene**/
             }else{
                 fprintf(stderr,"Error: not enough fields on line number %d:\n%s\n",g,sc);
                 exit(EXIT_FAILURE);
@@ -141,7 +141,7 @@ void ReadUMIcountMatrix(string in_file, double **n_c, double *N_c, double *n, st
             /****go to the next in the list of words****/
             token = strtok(NULL," \t,");
             if(token != NULL){
-                n_c_tmp[c] = atoi(token);
+                n_c_tmp[c] = stod(token);
                 n_tmp += n_c_tmp[c];// total count this gene
                 N_c[c] += n_c_tmp[c];// total count for this cell
             }else{
@@ -193,10 +193,10 @@ void Get_G_C_MTX(string in_file, int &N_rows, int &G, int &C, map<int,int> &gene
 		}else{
 			token = strtok(retval," ");
 			// Highest number of genes expressed and not
-			N_rows = atoi(token);
+			N_rows = stoi(token);
 			token = strtok(NULL," \t");
 			// Number of cells
-			C = atoi(token);
+			C = stoi(token);
 			break;
 		}
 	}
@@ -212,7 +212,7 @@ void Get_G_C_MTX(string in_file, int &N_rows, int &G, int &C, map<int,int> &gene
     while ( (retval = fgets(ss,1024,infp) ) != NULL) {
 		// Read values gene idx and add as expressed
 		token = strtok(retval," \t");
-		g = atoi(token);
+		g = stoi(token);
 		g=g-1;
 		// Update gene_idx map and count gene if not seen before
 		if( ! expressed_genes[g] ){
@@ -324,11 +324,11 @@ void ReadMTX(string mtx_file, string gene_name_file, string cell_name_file, doub
     while ( (retval = fgets(ss,1024,infp) ) != NULL) {
 		// Read values gene idx and add as expressed
 		token = strtok(retval," ");
-		g = atoi(token);
+		g = stoi(token);
 		token = strtok(NULL," ");
-		c = atoi(token);
+		c = stoi(token);
 		token = strtok(NULL," ");
-		count = atoi(token);
+		count = stod(token);
 
 		g = g-1;
 		c = c-1;
